requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
mutagen==1.47.0
pydantic==2.8.2
pydantic_core==2.20.1
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
PyYAML==6.0.2
sniffio==1.3.1
soundcloud-lib==0.6.1
soupsieve==2.6
starlette==0.38.2
typing_extensions==4.12.2
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.20.0
